requirements.txt
asgiref==3.7.2
Django==4.2.2
Pillow==9.5.0
sqlparse==0.4.4
typing-extensions==4.6.3
psycopg2-binary==2.9.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==5.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
django-rosetta==0.9.9
celery==5.3.1
redis==4.5.5
django-modeltranslation==0.18.11
uWSGI==2.0.20